+++ packages/orq-rc/src/orq_ai_sdk/models/bulkfileuploadop.py
@@ -81,5 +81,5 @@ class ResponseBody(BaseModel):
 
     file_name: str
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2024-11-26T10:00:21.420Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2024-12-01T21:30:45.101Z")
     r"""The date and time the resource was created"""