+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentinvokeop.py
@@ -492,6 +492,7 @@ DeploymentInvokeProvider = Literal[
     "fal",
     "leonardoai",
     "nvidia",
+    "jina",
 ]
 r"""The provider used to generate the response"""