+++ packages/orq-rc/src/orq_ai_sdk/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "2.0"
-    sdk_version: str = "1.31.0-rc.7"
+    sdk_version: str = "1.31.0-rc.12"
     gen_version: str = "2.467.4"
-    user_agent: str = "speakeasy-sdk/python 1.31.0-rc.7 2.467.4 2.0 orq-ai-sdk"
+    user_agent: str = "speakeasy-sdk/python 1.31.0-rc.12 2.467.4 2.0 orq-ai-sdk"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None