+++ packages/orq-rc/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v1.31.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v1.31.0-rc.7] https://pypi.org/project/orq-ai-sdk/1.31.0-rc.7 - packages/orq-rc
+
+## 2024-12-02 21:55:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.449.0 (2.467.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.31.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v1.31.0-rc.12] https://pypi.org/project/orq-ai-sdk/1.31.0-rc.12 - packages/orq-rc